@@ -6,16 +6,21 @@ class Gaps {
   int bottom;
   int left;
 
-  Gaps({this.inner, this.outer, this.top, this.right, this.bottom, this.left});
+  Gaps(
+      {required this.inner,
+      required this.outer,
+      required this.top,
+      required this.right,
+      required this.bottom,
+      required this.left});
 
-  Gaps.fromJson(Map<String, dynamic> json) {
-    inner = json['inner'];
-    outer = json['outer'];
-    top = json['top'];
-    right = json['right'];
-    bottom = json['bottom'];
-    left = json['left'];
-  }
+  Gaps.fromJson(Map<String, dynamic> json)
+      : inner = json['inner'],
+        outer = json['outer'],
+        top = json['top'],
+        right = json['right'],
+        bottom = json['bottom'],
+        left = json['left'];
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
