@@ -4,14 +4,17 @@ class Rect {
   int width;
   int height;
 
-  Rect({this.x, this.y, this.width, this.height});
+  Rect(
+      {required this.x,
+      required this.y,
+      required this.width,
+      required this.height});
 
-  Rect.fromJson(Map<String, dynamic> json) {
-    x = json['x'];
-    y = json['y'];
-    width = json['width'];
-    height = json['height'];
-  }
+  Rect.fromJson(Map<String, dynamic> json)
+      : x = json['x'],
+        y = json['y'],
+        width = json['width'],
+        height = json['height'];
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
