@@ -1,11 +1,10 @@
 class BarConfigEvent {
   String config;
 
-  BarConfigEvent({this.config});
+  BarConfigEvent({required this.config});
 
-  BarConfigEvent.fromJson(Map<String, dynamic> json) {
-    config = json.toString();
-  }
+  BarConfigEvent.fromJson(Map<String, dynamic> json)
+      : config = json['config'].toString();
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
