@@ -1,11 +1,9 @@
 class OutputEvent {
   String change;
 
-  OutputEvent({this.change});
+  OutputEvent({required this.change});
 
-  OutputEvent.fromJson(Map<String, dynamic> json) {
-    change = json['change'];
-  }
+  OutputEvent.fromJson(Map<String, dynamic> json) : change = json['change'];
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
