@@ -1,11 +1,9 @@
 class ShutdownEvent {
   String change;
 
-  ShutdownEvent({this.change});
+  ShutdownEvent({required this.change});
 
-  ShutdownEvent.fromJson(Map<String, dynamic> json) {
-    change = json['change'];
-  }
+  ShutdownEvent.fromJson(Map<String, dynamic> json) : change = json['change'];
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
