@@ -2,12 +2,11 @@ class TickEvent {
   bool first;
   String payload;
 
-  TickEvent({this.first, this.payload});
+  TickEvent({required this.first, required this.payload});
 
-  TickEvent.fromJson(Map<String, dynamic> json) {
-    first = json['first'];
-    payload = json['payload'];
-  }
+  TickEvent.fromJson(Map<String, dynamic> json)
+      : first = json['first'],
+        payload = json['payload'];
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
